move ray rules
