+++ eva/optimizer/rules/rules.py
@@ -20,6 +20,8 @@ from eva.catalog.catalog_manager import CatalogManager
 from eva.catalog.catalog_type import TableType
 from eva.catalog.catalog_utils import is_video_table
 from eva.constants import CACHEABLE_UDFS
+from eva.executor.execution_context import Context
+from eva.experimental.ray.planner.exchange_plan import ExchangePlan
 from eva.expression.expression_utils import (
     conjunction_list_to_expression_tree,
     to_conjunction_list,
@@ -66,6 +68,7 @@ from eva.optimizer.operators import (
     LogicalGroupBy,
     LogicalInsert,
     LogicalJoin,
+    LogicalExchange,
     LogicalLimit,
     LogicalLoadData,
     LogicalOrderBy,
@@ -478,6 +481,90 @@ class CombineSimilarityOrderByAndLimitToFaissIndexScan(Rule):
         for child in orderby_node.children:
             faiss_index_scan_node.append_child(child)
         yield faiss_index_scan_node
+
+class LogicalExchangeToPhysical(Rule):
+    def __init__(self):
+        pattern = Pattern(OperatorType.LOGICALEXCHANGE)
+        pattern.append_child(Pattern(OperatorType.DUMMY))
+        super().__init__(RuleType.LOGICAL_EXCHANGE_TO_PHYSICAL, pattern)
+
+    def promise(self):
+        return Promise.LOGICAL_EXCHANGE_TO_PHYSICAL
+
+    def check(self, grp_id: int, context: OptimizerContext):
+        return True
+
+    def apply(self, before: LogicalExchange, context: OptimizerContext):
+        after = ExchangePlan(before.view)
+        for child in before.children:
+            after.append_child(child)
+        yield after
+
+
+class LogicalProjectToPhysical(Rule):
+    def __init__(self):
+        pattern = Pattern(OperatorType.LOGICALPROJECT)
+        pattern.append_child(Pattern(OperatorType.DUMMY))
+        super().__init__(RuleType.LOGICAL_PROJECT_TO_PHYSICAL, pattern)
+
+    def promise(self):
+        return Promise.LOGICAL_PROJECT_TO_PHYSICAL
+
+    def check(self, grp_id: int, context: OptimizerContext):
+        return True
+
+    def apply(self, before: LogicalProject, context: OptimizerContext):
+        after = ProjectPlan(before.target_list)
+        for child in before.children:
+            after.append_child(child)
+        upper = ExchangePlan(
+            parallelism=1,
+            ray_conf={"num_gpus": 1} if Context().gpus else {"num_cpus": 1},
+        )
+        upper.append_child(after)
+        yield upper
+
+
+class LogicalGetToSeqScan(Rule):
+    def __init__(self):
+        pattern = Pattern(OperatorType.LOGICALGET)
+        super().__init__(RuleType.LOGICAL_GET_TO_SEQSCAN, pattern)
+
+    def promise(self):
+        return Promise.LOGICAL_GET_TO_SEQSCAN
+
+    def check(self, before: LogicalGet, context: OptimizerContext):
+        return True
+
+    def apply(self, before: LogicalGet, context: OptimizerContext):
+        # Configure the batch_mem_size. It decides the number of rows
+        # read in a batch from storage engine.
+        # ToDO: Experiment heuristics.
+        scan = SeqScanPlan(None, before.target_list, before.alias)
+        lower = ExchangePlan(parallelism=1)
+        lower.append_child(
+            StoragePlan(
+                before.table_obj,
+                before.video,
+                predicate=before.predicate,
+                sampling_rate=before.sampling_rate,
+            )
+        )
+        scan.append_child(lower)
+        # Check whether the projection contains a UDF
+        if before.target_list is None or not any(
+            [isinstance(expr, FunctionExpression) for expr in before.target_list]
+        ):
+            yield scan
+        else:
+            upper = ExchangePlan(
+                parallelism=2,
+                ray_conf={"num_gpus": 1}
+                if len(Context().gpus) >= 2
+                else {"num_cpus": 1},
+            )
+            upper.append_child(scan)
+            yield upper
 
 
 # REWRITE RULES END
